Add files via upload
You need to use the input files that were shared with assignment. You need to  complete the assignment to convert the input to xml format. 

Certain things to be careful when writing the code :

Go through the assignment in detail and try to deliver a professional developed solution with proper instructions like readme file, input files and output file.Solution should be executableAdd proper comments in source codeCode should be neat, easy to understand and bug-freeThe better the code solutions, more chances for interview round else strict NO.

Please complete the code test and  send it to me by  21st  Noon .